+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Released 2021-03-XX
 - Added `config.ini` and `censearch_config` to set/parse the Twitter API Keys
 - Search terms are now displayed before running the search
 - Each tweet displayed is underlined to make it easier to visually separate
+- Option to output found tweets to a specified file.
 
 ## Version 0.1.1